Refactor runtime driver and HAP internals
